change the precision of T in forsythe

+++ test/test_forsythe.jl
@@ -2,7 +2,7 @@ n = rand(1:10)
 @test matrixdepot("forsythe", n) == 
                 matrixdepot("forsythe", n, sqrt(eps(Float64)), zero(Float64))
 @test matrixdepot("forsythe", n , 1, 2) == matrixdepot("forsythe", Float64, n, 1, 2)
-T = Float16
+T = Float32
 a = convert(T, 3)
 b = convert(T, 4)
 @test matrixdepot("forsythe", T, n, 3, 4) == matrixdepot("forsythe", n, a, b)
